Add command !rescue to unlock a claim for a limited time.
The !rescue command is written in squad chat (only) by the claim holding SL.
The vehicle type he has claim for is then unlocked for a limited time.

Both the claim holding SL and the rescuer is informed when the time
expires.

The SL can at any time restart/extend the rescue period by running the
command again.