vehicle-claims: bug fix: Remove all dead squad claims
Don't just remove old claims from the just-created squad. Go through all
claims and remove any that belong to squads no longer existing.

This is a workaround for the fact that the server doesn't tell us when a
squad is disbanded, so we have to figure that out when a new squad is created.

Also: Use faction name instead of team number in some log lines.