vehicle-claims: Fix issue where CRF is selectable but unlisted
There is apparently a mismatch between the SDK and the actual game.
The SDK says CRF is not available on Manicouagan_RAAS_v2, but it is
available to play in-game.

Work around this by using the specified faction and tacking on the default
unit from a listed faction on the same team.